Fix test skipping on py3.5 or lower.
